packages/virion-labs-business-logic-api/requirements.txt
fastapi
pyjwt>=2.8.0
uvicorn
httpx
python-dotenv
pydantic-settings
pydantic[email]
python-multipart